The English - old Norse dictionary in its original form as found on:
http://lexicon.ff.cuni.cz/txt/oi_cleasbyvigfusson.txt

and a perl script to convert it to proper UTF-8 encoded, valid XML.

The XML is very bare-bones, and does not yet give satisfactory results (ie all articles run together in one long paragraph, with only some headings in between splitting them up).

The XML has an xhtml outer wrapper, to make the output readable for standard web browsers, but this can easily (and should) be changed.